+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [2.0.0]
  - Rename `parquetMetadataAsync` to `parquetMetadata`
  - Previous `parquetMetadata` is now `import { parquetMetadataSync } from 'hyparquet/src/metadata.js'`
+ - Rename `onPage.columnName: string` to `onPage.pathInSchema: string[]`
+ - Remove `rowFormat` and always return rows as objects
 
 ## [1.23.0]
  - Replace `columnName: string` with `pathInSchema: string[]` in `onPage` callback (#144)